Check task data before associate it to database

+++ task_service.php
@@ -47,38 +47,50 @@
     }
 
     public function markAsDone($task) {
-      $sql = 'UPDATE tb_task SET done = 1 WHERE id = :id AND id_user = :id_user;';
-      $pdo = $this->connection->connect();
-      $stmt = $pdo->prepare($sql);
-      $stmt->bindValue(':id', $task->__get('id'));
-      $stmt->bindValue(':id_user', $task->__get('id_user'));
-
-      return $stmt->execute();
+      if($task->id && $task->id_user){
+        $sql = 'UPDATE tb_task SET done = 1 WHERE id = :id AND id_user = :id_user;';
+        $pdo = $this->connection->connect();
+        $stmt = $pdo->prepare($sql);
+        $stmt->bindValue(':id', $task->__get('id'));
+        $stmt->bindValue(':id_user', $task->__get('id_user'));
+        
+        if($stmt->execute()) {
+          return $stmt->rowCount();
+        }
+      }
+      return false;
     }
 
     public function deleteTask($task) {
-      $sql = 'DELETE FROM tb_task WHERE id = :id AND id_user = :id_user';
-      $pdo = $this->connection->connect();
-      $stmt = $pdo->prepare($sql);
-      $stmt->bindValue(':id', $task->__get('id'));
-      $stmt->bindValue(':id_user', $task->__get('id_user'));
-
-      return $stmt->execute();
+      if($task->id && $task->id_user) {
+        $sql = 'DELETE FROM tb_task WHERE id = :id AND id_user = :id_user';
+        $pdo = $this->connection->connect();
+        $stmt = $pdo->prepare($sql);
+        $stmt->bindValue(':id', $task->__get('id'));
+        $stmt->bindValue(':id_user', $task->__get('id_user'));
+        if($stmt->execute()) {
+          return $stmt->rowCount();
+        }
+      }
+      return false;
     }
 
     public function updateTask($task) {
-      $sql = '
-        UPDATE tb_task SET task = ?, task_description = ?
-        WHERE id_user = ? AND id = ?
-      ';
-      $pdo = $this->connection->connect();
-      $stmt = $pdo->prepare($sql);
-      $stmt->bindValue(1, $task->__get('task'));
-      $stmt->bindValue(2, $task->__get('description'));
-      $stmt->bindValue(3, $task->__get('id_user'));
-      $stmt->bindValue(4, $task->__get('id'));
-
-      return $stmt->execute();
+      if($task->task && $task->task_description && $task->id_user && $task->id) {
+        $sql = '
+          UPDATE tb_task SET task = ?, task_description = ?
+          WHERE id_user = ? AND id = ?
+        ';
+        $pdo = $this->connection->connect();
+        $stmt = $pdo->prepare($sql);
+        $stmt->bindValue(1, $task->__get('task'));
+        $stmt->bindValue(2, $task->__get('description'));
+        $stmt->bindValue(3, $task->__get('id_user'));
+        $stmt->bindValue(4, $task->__get('id'));
+        
+        return $stmt->execute();
+      }
+      return false;
     }
 
     public function preventHTMLInjection($task) {
